+++ src/lib/modus/simulation/resource_system.ex
@@ -32,7 +32,7 @@ defmodule Modus.Simulation.ResourceSystem do
 
   @impl true
   def init(state) do
-    Phoenix.PubSub.subscribe(Modus.PubSub, "simulation:ticks")
+    Phoenix.PubSub.subscribe(Modus.PubSub, "modus:tick")
     {:ok, state}
   end