+++ src/lib/modus/simulation/seasons.ex
@@ -114,7 +114,7 @@ defmodule Modus.Simulation.Seasons do
 
   @impl true
   def init(state) do
-    Phoenix.PubSub.subscribe(Modus.PubSub, "simulation:ticks")
+    Phoenix.PubSub.subscribe(Modus.PubSub, "modus:tick")
     {:ok, state}
   end